+++ src/components/map/popups/MapPopupAirport.vue
@@ -93,6 +93,27 @@
             </div>
             <airport-aircraft/>
         </template>
+        <template #actions>
+            <common-button-group>
+                <common-button :to="`/airport/${ airport.icao }`">
+                    <template #icon>
+                        <data-icon/>
+                    </template>
+                    Dashboard
+                </common-button>
+                <common-button @click="copy.copy(`${ config.public.DOMAIN }/?airport=${ airport.icao }`)">
+                    <template #icon>
+                        <share-icon/>
+                    </template>
+                    <template v-if="copy.copyState.value">
+                        Copied!
+                    </template>
+                    <template v-else>
+                        Link
+                    </template>
+                </common-button>
+            </common-button-group>
+        </template>
     </common-info-popup>
 </template>
 
@@ -118,6 +139,10 @@ import CommonToggle from '~/components/common/basic/CommonToggle.vue';
 import AirportInfo from '~/components/views/airport/AirportInfo.vue';
 import AirportAircraft from '~/components/views/airport/AirportAircraft.vue';
 import AirportControllers from '~/components/views/airport/AirportControllers.vue';
+import CommonButtonGroup from '~/components/common/basic/CommonButtonGroup.vue';
+import CommonButton from '~/components/common/basic/CommonButton.vue';
+import DataIcon from '@/assets/icons/kit/data.svg?component';
+import ShareIcon from '@/assets/icons/kit/share.svg?component';
 
 const props = defineProps({
     overlay: {
@@ -134,6 +159,8 @@ const aircraft = getAircraftForAirport(overlayData);
 const store = useStore();
 const mapStore = useMapStore();
 const dataStore = useDataStore();
+const copy = useCopyText();
+const config = useRuntimeConfig();
 
 const airport = computed(() => dataStore.vatspy.value?.data.airports.find(x => x.icao === props.overlay.data.icao));
 const vatAirport = computed(() => dataStore.vatsim.data.airports.value.find(x => x.icao === props.overlay.data.icao));
@@ -211,6 +238,14 @@ const tabs = computed<InfoPopupContent>(() => {
     }
 
     if (!list.info.sections.length) delete list.info;
+
+    list.atc.sections.push({
+        key: 'actions',
+    });
+
+    list.info?.sections.push({
+        key: 'actions',
+    });
 
     return list;
 });